% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{check_changepoint_input}
\alias{check_changepoint_input}
\title{Check changepoint Input

A generic function.}
\usage{
check_changepoint_input(changepoint, ndf, pdim, verbose = TRUE)
}
\arguments{
\item{changepoint}{estimated changepoint in dimension function}

\item{ndf}{The number of degrees of freedom of x.}

\item{pdim}{The number of dimensions of x.}

\item{verbose}{output message}
}
\description{
Check changepoint Input

A generic function.
}
