% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{check_comp_input}
\alias{check_comp_input}
\title{Check components Input

A generic function.}
\usage{
check_comp_input(components, ndf, pdim, verbose = FALSE)
}
\arguments{
\item{components}{A series of right singular vectors to estimate.
Components must be smaller or equal to min(nrow(x),ncol(x)).}

\item{ndf}{The number of degrees of freedom of x.}

\item{pdim}{The number of dimensions of x.}

\item{verbose}{output message}
}
\description{
Check components Input

A generic function.
}
