% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dim_matrix.R
\name{check_dim_matrix}
\alias{check_dim_matrix}
\title{Return parameter setting for subspace function}
\usage{
check_dim_matrix(x, rnk = NA, verbose = FALSE)
}
\arguments{
\item{x}{A numeric real-valued matrix with n number of samples and
p number of features.
If p>n, a warning message is generated and the transpose of x is used.}

\item{rnk}{The number of right singular vectors to estimate.
rnk must be smaller or equal to max(nrow(x),ncol(x)).}

\item{verbose}{output message}
}
\value{
Returns a list with entries:
\describe{
  \item{ndf:}{ The number of degrees of freedom of x.}
  \item{pdim:}{ The number of dimensions of x.}
  \item{svr:}{ The samples to variance ratio: equals to ndf divided by pdim.}
  \item{rnk:}{ The number of right singular vectors estimated.}
  \item{transpose_flag:}{ A logical value indicating whether the
  matrix x is transposed.}
}
}
\description{
Return parameter setting for subspace function
}
\examples{
x <- x_sim(n = 100, p = 150, ncc = 10, var = c(rep(10, 5), rep(1, 5)))
params <- check_dim_matrix(x, rnk = 40)
}
\seealso{
[checkDesignMatrix()] from Random Matrix Theory pacakge
credit to Gregory Giecold and Lionel Ouaknin.
}
