% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{check_num_est_samples_input}
\alias{check_num_est_samples_input}
\title{Check num_est_samples Input

A generic function.}
\usage{
check_num_est_samples_input(num_est_samples, ndf, pdim, verbose = TRUE)
}
\arguments{
\item{num_est_samples}{Split data into num_est_samples-fold
for parallel computation.}

\item{ndf}{The number of degrees of freedom of x.}

\item{pdim}{The number of dimensions of x.}

\item{verbose}{output message}
}
\description{
Check num_est_samples Input

A generic function.
}
