% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{check_times_input}
\alias{check_times_input}
\title{Check Times Input

A generic function.}
\usage{
check_times_input(times, ndf, pdim, verbose = TRUE)
}
\arguments{
\item{times}{Split data into times-fold for parallel computation.}

\item{ndf}{The number of degrees of freedom of x.}

\item{pdim}{The number of dimensions of x.}

\item{verbose}{output message}
}
\description{
Check Times Input

A generic function.
}
