% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDimMatrix.R
\name{CheckDimMatrix}
\alias{CheckDimMatrix}
\title{Return parameter setting for subspace function}
\usage{
CheckDimMatrix(X, rnk = NA, verbose = TRUE)
}
\arguments{
\item{X}{A numeric real-valued matrix with n number of samples and p number of features. If p>n, a warning message is generated and the transpose of X is used.}

\item{rnk}{The number of right singular vectors to estimate. rnk must be smaller or equal to max(nrow(X),ncol(X)).}
}
\value{
Returns a list with entries:
\describe{
  \item{ndf:}{ The number of degrees of freedom of X.}
  \item{pdim:}{ The number of dimensions of X.}
  \item{svr:}{ The samples to variance ratio: equals to ndf divided by pdim.}
  \item{rnk:}{ The number of right singular vectors estimated.}
  \item{transpose_flag:}{ A logical value indicating whether the matrix X is transposed.}
}
}
\description{
Return parameter setting for subspace function
}
\examples{
\donttest{
X <- Xsim(n = 150, p = 100, ncc = 10, var = 2)
params <- CheckDimMatrix(X, rnk = 40)
}

}
\seealso{
[checkDesignMatrix()] from Random Matrix Theory pacakge credit to Gregory Giecold and Lionel Ouaknin.
}
