% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{CheckRankInput}
\alias{CheckRankInput}
\title{Check Rank Input

A generic function.}
\usage{
CheckRankInput(rank, ndf, pdim, verbose = TRUE)
}
\arguments{
\item{rank}{A series of right singular vectors to estimate. rank must be smaller or equal to min(nrow(X),ncol(X)).}

\item{ndf}{The number of degrees of freedom of X.}

\item{pdim}{The number of dimensions of X.}

\item{verbose}{output message}
}
\description{
Check Rank Input

A generic function.
}
