% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{CheckTimesInput}
\alias{CheckTimesInput}
\title{Check Times Input

A generic function.}
\usage{
CheckTimesInput(times, ndf, pdim, verbose = TRUE)
}
\arguments{
\item{times}{Split data into X times for parallel computation.}

\item{ndf}{The number of degrees of freedom of X.}

\item{pdim}{The number of dimensions of X.}

\item{verbose}{output message}
}
\description{
Check Times Input

A generic function.
}
