% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_eigenvalues.R
\name{correct_eigenvalues}
\alias{correct_eigenvalues}
\title{correct eigenvalues from a subspace object and return
a subspace object}
\usage{
correct_eigenvalues(subspace, num_est_samples, verbose, ...)
}
\arguments{
\item{num_est_samples}{the number of resamples to take from the
Marcenko-Pastur distribution to estimate the eigenvalues.}

\item{verbose}{output message}

\item{...}{Extra parameters}

\item{x}{A subspace object.}
}
\value{
Returns a list with entries:
\describe{
  \item{ndf:}{ The number of degrees of freedom of x.}
  \item{pdim:}{ The number of dimensions of x.}
  \item{components:}{ A series of right singular vectors estimated.}
  \item{var_correct:}{ Corrected population variance
   for Marcenko-Pastur distribution.}
  \item{transpose_flag:}{ A logical value indicating
   whether the matrix x is transposed.}
  \item{irl:}{ A data frame of scaled eigenvalues for
   specified components and corresponding dimensions.}
  \item{sigma_a:}{ A vector of corrected eigenvalues up to max(components).}
  \item{mp_irl:}{ A data frame of sampled expected eigenvalues from
   Marcenko-Pastur for specified components and corresponding dimensions.}
  \item{sigma_mp:}{ A vector of samped expected eigenvalues from
   Marcenko-Pastur up to max(components).}
  \item{v:}{ Right singular vectors of x matrix for specified components.}
  \item{u:}{ Left singular vectors of x matrix or specified components.}
}
}
\description{
This function correct eigenvalues from a subspace by
substracting sampling eigenvalues from a random noise matrix N of
the same dimension as x,which follows a Marcenko-Pastur distribution
with package "RMTsata
"(https://cran.r-project.org/web/packages/RMTstat/index.html).
}
\examples{
x <- x_sim(n = 100, p = 150, ncc = 10, var = c(rep(10, 5), rep(1, 5)))
x \%>\% create_subspace(components = 8:30) \%>\% correct_eigenvalues() \%>\% plot()
}
\seealso{
* [MarchenkoPasturPar()] calculates upper and lower limits
 of Marcenko-Pastur distribution from RMTstat package.

* [rmp()] sample scaled eigenvalues of random noise matrix
 from RMTstat package.
}
