% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subspace.R
\name{create_subspace}
\alias{create_subspace}
\title{A constructor function for the subspace class}
\usage{
create_subspace(
  x,
  components = NULL,
  decomposition = c("svd", "eigen"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric real-valued matrix with n number of samples and
p number of features. If p > n, a warning message is generated and
the transpose of x is used.}

\item{components}{A series of right singular vectors to estimate.
Components must be smaller or equal to min(nrow(x), ncol(x)).}

\item{decomposition}{The method to be used; method = "svd"
returns results from singular value decomposition; method = "eigen"
returns results from eigenvalue decomposition.}

\item{verbose}{output message}

\item{...}{Extra parameters}
}
\value{
Returns a list with entries:
\describe{
  \item{ndf:}{ The number of degrees of freedom of x.}
  \item{pdim:}{ The number of dimensions of x.}
  \item{components:}{ A series of right singular vectors estimated.}
  \item{transpose_flag:}{ A logical value indicating
   whether the matrix x is transposed.}
  \item{irl:}{ A data frame of scaled eigenvalues for
   specified components and corresponding dimensions.}
  \item{sigma_a:}{ A vector of scaled eigenvalues up to max(components).}
  \item{v:}{ Right singular vectors of x matrix for specified components.}
  \item{u:}{ Left singular vectors of x matrix or specified components.}
}
}
\description{
This function calculates scaled eigenvalues
 and eigenvectors of x matrix.
}
\examples{
x <- x_sim(n = 100, p = 150, ncc = 10, var = c(rep(10, 5), rep(1, 5)))
x \%>\% create_subspace(components = 8:30) \%>\% str()
}
