% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{dmp}
\alias{dmp}
\title{dmp within package "RMTsata".}
\usage{
dmp(x, ndf = NA, pdim = NA, var = 1, svr = ndf/pdim, log = FALSE)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{ndf}{The number of degrees of freedom of x.}

\item{pdim}{The number of dimensions of x.}

\item{var}{Population variance.}

\item{svr}{ndf/pdim.}

\item{log}{A logical value.}
}
\description{
All credit to Iain M. Johnstone, Zongming Ma,
 Patrick O. Perry and Morteza Shahram
 (https://cran.r-project.org/web/packages/RMTstat/index.html).
}
