% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rank_ladle.R
\name{estimate_rank_ladle}
\alias{estimate_rank_ladle}
\title{ladle}
\usage{
estimate_rank_ladle(x)
}
\arguments{
\item{x}{A numeric real-valued matrix with n number of samples and
p number of features (n > p).}

\item{nboot}{bootstrap sample size.}
}
\value{
Returns a list with entries:
\describe{
  \item{fn0:}{ fn0 is the bootstrap eigenvector variability.}
  \item{fn:}{ fn is the renormalized fn0.}
  \item{lam:}{ lam is the sample eigenvalues.}
  \item{phin:}{ phin is the renormalized and shifted sample eigenvalues.}
  \item{gn:}{ gn is the objective function ("y" in ladle plot).}
  \item{d:}{ d is the ladle estimator.}
}
}
\description{
All credit to Luo, W. and Li, B. (2016),
Combining Eigenvalues and Variation of Eigenvectors for Order
Determination, Biometrika, 103, 875–887. <doi:10.1093/biomet/asw051>
}
\examples{
x <- x_sim(n = 100, p = 150, ncc = 10, var = c(rep(10, 5), rep(1, 5)))
x \%>\% estimate_rank_ladle()
}
