% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rank_posterior_cor.R
\name{estimate_rank_posterior_cor}
\alias{estimate_rank_posterior_cor}
\title{Signal subspace dimension estimation in high-dimensional matrix by bcp posterior}
\usage{
estimate_rank_posterior_cor(subspace_ = NULL, verbose = TRUE, ...)
}
\arguments{
\item{subspace_}{A subspace class.}

\item{verbose}{output message}

\item{...}{Extra parameters}
}
\value{
Returns a list with entries:
\describe{
  \item{ndf:}{ The number of degrees of freedom of x.}
  \item{pdim:}{ The number of dimensions of x.}
  \item{components:}{ A series of right singular
   vectors estimated.}
  \item{var_correct:}{ Corrected population variance
   for Marchenko-Pastur distribution.}
  \item{transpose_flag:}{ A logical value indicating
   whether the matrix x is transposed.}
  \item{irl:}{ A data frame of scaled eigenvalues for
   specified rank and corresponding dimensions.}
  \item{mp_irl:}{ A data frame of sampled expected eigenvalues from
   Marchenko-Pastur for specified rank and corresponding dimensions.}
  \item{v:}{ Right singular vectors of x matrix for specified rank.}
  \item{u:}{ Left singular vectors of x matrix or specified rank.}
  \item{dimension:}{ Estimated signal subspace dimension.}
  \item{bcp_irl:}{ Probability of change in mean and posterior means
   of eigenvalue difference between $x$ and $N$.}
}
}
\description{
Estimate the dimension of a signal-rich subspace in large,
 high-dimensional data.
}
\section{Details}{

 We estimate the intrinsic dimension of a signal-rich subspace
 in large high-dimensional data by decomposing matrix into a
 signal-plus-noise space and approximate the signal-rich subspace
 with a rank K approximation \eqn{\hat{x}=\sum_{k=1}^{K}d_ku_k{v_k}^T}.
 To estimate rank K, we propose a simple procedure assuming that matrix
 x is composed of a low-rank signal matrix S and an average general noise
 random matrix \eqn{\bar{N}}. It has been shown that the average eigenvalues
 of random matrices N follows a universal Marchenko-Pastur (MP) distribution.
 We hypothesize that the deviation of eigenvalues of x from the MP
 distribution indicates the intrinsic dimension of signal-rich subspace.
}

\examples{
x <- x_sim(n = 100, p = 150, ncc = 10, var = c(rep(10, 5), rep(1, 5)))
results <- x \%>\%
create_subspace(components = 8:30) \%>\%
correct_eigenvalues() \%>\%
estimate_rank_posterior_cor()

str(results)
plot(results$Subspace, changepoint = results$dimension,
     annotation = 10)
modified_legacyplot(results$bcp_irl, annotation = 10)
}
\seealso{
[RMTstat] for details of Marchenko-Pastur distribution.

https://dracodoc.wordpress.com/2014/07/21/
a-simple-algorithm-to-detect-flat-segments-in-noisy-signals/ for detection
of flat and spike in noisy signals
}
