% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{gs1}
\alias{gs1}
\title{Gram-schmidt process}
\usage{
gs1(x)
}
\arguments{
\item{X}{a numeric matrix.}
}
\description{
This function uses Gram-schmidt process to generate orthorgonal basis for a vector space.
}
\examples{
\donttest{
# 3 baskets, each with enrollement size 5
X <- Xsim(n=100,p=90,ncc=2,var=10)
gs1(X)
}
}
