% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{hoff}
\alias{hoff}
\title{hoff}
\usage{
hoff(y = y)
}
\arguments{
\item{y}{A numeric real-valued matrix with n number of samples and
p number of features (n > p).}
}
\description{
All credit to Luo, W. and Li, B. (2016),
Combining Eigenvalues and Variation of Eigenvectors for Order
Determination, Biometrika, 103, 875–887. <doi:10.1093/biomet/asw051>
}
