% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{hoff_gibbs_sample}
\alias{hoff_gibbs_sample}
\title{hoff_gibbs_sample}
\usage{
hoff_gibbs_sample(e, nu, nv, phi, mu, psi, lor, llb = 50, lub = 2000)
}
\arguments{
\item{e}{An R object.}

\item{nu}{An R object.}

\item{nv}{An R object.}

\item{phi}{An R object.}

\item{mu}{An R object.}

\item{psi}{An R object.}

\item{lor}{An R object.}

\item{llb}{An R object.}

\item{lub}{An R object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
