% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{hoff_gibbs_UVD_fixedrank}
\alias{hoff_gibbs_UVD_fixedrank}
\title{hoff_gibbs_UVD_fixedrank}
\usage{
hoff_gibbs_UVD_fixedrank(U, V, D, Y, phi, mu, psi)
}
\arguments{
\item{U}{An R object.}

\item{V}{An R object.}

\item{D}{An R object.}

\item{Y}{An R object.}

\item{phi}{An R object.}

\item{mu}{An R object.}

\item{psi}{An R object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
