% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{hoff_rmvnorm}
\alias{hoff_rmvnorm}
\title{hoff_rmvnorm}
\usage{
hoff_rmvnorm(mu, Sig2)
}
\arguments{
\item{mu}{An \tt{R} object.}

\item{Sig2}{An \tt{R} object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
