% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_k_means.R
\name{KM}
\alias{KM}
\title{Modified K means clustering}
\usage{
KM(X, k = 2)
}
\arguments{
\item{X}{an n by p numeric matrix; the data
matrix of predictors}

\item{k}{integer giving the number of neighbors to
include. k = 2.}
}
\value{
a data frame of total within-cluster variation.
}
\description{
k-means clustering modified to cluster probabilities of change points
 to 2 clusters and return total within-cluster variation.
}
