% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_k_means.R
\name{km_plot}
\alias{km_plot}
\title{Summary plots of km() output}
\usage{
km_plot(x, annotation = NULL)
}
\arguments{
\item{x}{A data frame. The result of a call to km().}

\item{annotation}{Annotate points up to total components calculated.
No annotation when annotation = 0.}
}
\description{
This function produces summary plots of km() output.
}
