% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{labc}
\alias{labc}
\title{labc}
\usage{
labc(theta, lmax, m, E2, phi, mu, psi)
}
\arguments{
\item{theta}{An \tt{R} object.}

\item{lmax}{An \tt{R} object.}

\item{m}{An \tt{R} object.}

\item{E2}{An \tt{R} object.}

\item{phi}{An \tt{R} object.}

\item{mu}{An \tt{R} object.}

\item{psi}{An \tt{R} object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
