% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_estimator_code.R
\name{ladle}
\alias{ladle}
\title{ladle}
\usage{
ladle(
  x = x,
  y = y,
  nslices = nslices,
  nboot = nboot,
  method = method,
  order = order
)
}
\arguments{
\item{x}{A numeric real-valued matrix with n number of samples and
p number of features (n > p).}

\item{y}{A vector.}

\item{nslices}{nslices is needed for sufficient dimension reduction.}

\item{nboot}{bootstrap sample size.}

\item{method}{options for method include: "pca", "cca", "ica", "sir",}

\item{order}{the order of the polynomial in polynomial}
}
\value{
Returns a list with entries:
\describe{
  \item{fn0:}{ fn0 is the bootstrap eigenvector variability.}
  \item{fn:}{ fn is the renormalized fn0.}
  \item{lam:}{ lam is the sample eigenvalues.}
  \item{phin:}{ phin is the renormalized and shifted sample eigenvalues.}
  \item{gn:}{ gn is the objective function ("y" in ladle plot).}
  \item{d:}{ d is the ladle estimator.}
}
}
\description{
All credit to Luo, W. and Li, B. (2016),
Combining Eigenvalues and Variation of Eigenvectors for Order
Determination, Biometrika, 103, 875–887. <doi:10.1093/biomet/asw051>
}
