% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{ldmvnorm}
\alias{ldmvnorm}
\title{ldmvnorm}
\usage{
ldmvnorm(y, mu, Sig)
}
\arguments{
\item{y}{An \tt{R} object.}

\item{mu}{An \tt{R} object.}

\item{Sig}{An \tt{R} object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
