% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacyplot.R
\name{legacyplot}
\alias{legacyplot}
\title{Summary plots of bcp() output}
\usage{
legacyplot(x, annotation = NULL, medianfilter = FALSE)
}
\arguments{
\item{x}{A list. The result of a call to bcp().}

\item{annotation}{Annotate points up to total components calculated.
No annotation when annotation = 0.}

\item{medianfilter}{A logical value. Compute running medians to smooth
scatter plot.}
}
\description{
This function produces summary plots of bcp() output.
 It is adapted from the default legacyplot function in pacakge bcp.
}
\examples{
library(bcp)
legacyplot(bcp(as.vector(c(rep(10, 10), 9.5, rep(0, 10))),
p0 = 0.1))
}
