% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{ln2moment}
\alias{ln2moment}
\title{ln2moment}
\usage{
ln2moment(mu, sigma, lmax)

ln2moment(mu, sigma, lmax)
}
\arguments{
\item{mu}{An R object.}

\item{sigma}{An R object.}

\item{lmax}{An R object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.

All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
