% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{lpY_binary}
\alias{lpY_binary}
\title{lpY_binary}
\usage{
lpY_binary(Y, Theta)
}
\arguments{
\item{Y}{An \tt{R} object.}

\item{Theta}{An \tt{R} object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
