% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung.R
\docType{data}
\name{lung}
\alias{lung}
\title{IPF Single Cell Altas data}
\format{An object of class Seurat.}
\source{
\href{https://www.biorxiv.org/content/10.1101/759902v1.full.pdf}
}
\usage{
data(lung)
}
\description{
The lung data is a subset of patient "001C" profiled in IPF
Single Cell Altas. The full data seet can be found on GEO (GSE136831),
which was examined  by
\href{https://www.biorxiv.org/content/10.1101/759902v1.full.pdf}
to build a single cell atlas of Idiopathic Pulmonary Fibrosis (IPF).
It contains the 60603 gene expression of 158 cells profiled from
"distal lung parenchyma samples obtained from 32 IPF, 18 chronic obstructive
pulmonary disease (COPD) and 29 control donor lungs" (Adamset al.2019).
}
\examples{
\dontrun{
data(lung)
}
}
\references{
Adamset  al.(2019),
(\href{https://www.biorxiv.org/content/10.1101/759902v1.full.pdf})
}
\keyword{datasets}
