% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_linear.R
\name{main_linear}
\alias{main_linear}
\title{Fit the linear exponential model using GenSA}
\usage{
main_linear(X, scale = TRUE, d_start = 10, rnk = 20, lower = c(-10,
  -1000, -100), upper = c(100, 1000, 100), plot = TRUE,
  times = 10000, seed = 1000)
}
\arguments{
\item{X}{a numeric matrix.}

\item{d_start}{the dimention start to fit the linear exponential model.}

\item{rnk}{number of right singular vectors to estimate.}

\item{lower}{Vector with length of par. Lower bounds for components.}

\item{upper}{Vector with length of par. Upper bounds for components.}

\item{plot}{a logical value. Ouput plots if true.}

\item{times}{Integer. Total simulation times for noise matrix.}

\item{seed}{the random number seed.}

\item{sacle}{a logical value.}
}
\description{
Fit the eigenvalues of X matrix with linear exponential model.
}
\examples{
\donttest{
# Simulate a matrix X with first two columns correlated.
X <- Xsim(n=100,p=90,ncc=2,var=10)


res <- main_linear(X,scale=TRUE,d_start=2,rnk=20)
}
}
