% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MarchenkoPasturPar}
\alias{MarchenkoPasturPar}
\title{Marcenko-Pastur distribution within package "RMTsata".}
\usage{
MarchenkoPasturPar(ndf = NA, pdim = NA, var = 1, svr = ndf/pdim)
}
\description{
All credit to Iain M. Johnstone, Zongming Ma, Patrick O. Perry and Morteza Shahram (https://cran.r-project.org/web/packages/RMTstat/index.html).
}
