% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_estimator_code.R
\name{matpower}
\alias{matpower}
\title{matpower}
\usage{
matpower(a, alpha)
}
\arguments{
\item{a}{an original matrix}

\item{alpha}{power}
}
\description{
All credit to Luo, W. and Li, B. (2016),
Combining Eigenvalues and Variation of Eigenvectors for Order
Determination, Biometrika, 103, 875–887. <doi:10.1093/biomet/asw051>
}
