% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_estimator_code.R
\name{mhat_cca}
\alias{mhat_cca}
\title{mhat_cca}
\usage{
mhat_cca(x, y)
}
\arguments{
\item{x}{numbers.}

\item{y}{numbers.}
}
\description{
All credit to Luo, W. and Li, B. (2016),
Combining Eigenvalues and Variation of Eigenvectors for Order
Determination, Biometrika, 103, 875–887. <doi:10.1093/biomet/asw051>
}
