% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_estimator_code.R
\name{mhat_ica}
\alias{mhat_ica}
\title{mhat_ica}
\usage{
mhat_ica(x)
}
\arguments{
\item{x}{numbers.}
}
\description{
All credit to Luo, W. and Li, B. (2016),
Combining Eigenvalues and Variation of Eigenvectors for Order
Determination, Biometrika, 103, 875–887. <doi:10.1093/biomet/asw051>
}
