% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_estimator_code.R
\name{mhat_ksir}
\alias{mhat_ksir}
\title{mhat_ksir}
\usage{
mhat_ksir(x = x, phi = phi, y = y, nslices = nslices)
}
\arguments{
\item{x}{numbers.}

\item{phi}{numbers.}

\item{y}{numbers.}

\item{nslices}{nslices is needed for sufficient dimension reduction.}
}
\description{
All credit to Luo, W. and Li, B. (2016),
Combining Eigenvalues and Variation of Eigenvectors for Order
Determination, Biometrika, 103, 875–887. <doi:10.1093/biomet/asw051>
}
