% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_legacyplot.R
\name{modified_legacyplot}
\alias{modified_legacyplot}
\title{Summary plots of bcp() output}
\usage{
modified_legacyplot(x, annotation = NULL, medianfilter = FALSE)
}
\arguments{
\item{annotation}{Annotate points up to total rank calculated. No annotation when annotation=0.}

\item{medianfilter}{A logical value. Compute running medians to smooth scatter plot.}

\item{X}{A list. The result of a call to bcp().}
}
\description{
This function produces summary plots of bcp() output. It is adapted from the default legacyplot function in pacakge bcp.
}
\examples{
\donttest{
modified_legacyplot(bcp(as.vector(c(rep(0,9),100,rep(0,10))),p0=0.1))
}
}
