% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{Null}
\alias{Null}
\title{Null}
\usage{
Null(M)
}
\arguments{
\item{M}{An R object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
