% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{pmp}
\alias{pmp}
\title{pmp within package "RMTsata".}
\usage{
pmp(q, ndf = NA, pdim = NA, var = 1, svr = ndf/pdim,
  lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{ndf}{The number of degrees of freedom of X.}

\item{pdim}{The number of dimensions of X.}

\item{var}{Population variance.}

\item{svr}{ndf/pdim.}

\item{lower.tail}{A logical value.}

\item{log.p}{A logical value.}
}
\description{
All credit to Iain M. Johnstone, Zongming Ma, Patrick O. Perry and Morteza Shahram (https://cran.r-project.org/web/packages/RMTstat/index.html).
}
