% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R
\name{print.dimension}
\alias{print.dimension}
\title{Print dimension

A generic function.}
\usage{
print.dimension(x, ...)
}
\arguments{
\item{x}{a dimension class.}

\item{...}{Extra parameters}
}
\description{
Print dimension

A generic function.
}
