% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{print.subspace}
\alias{print.subspace}
\title{Print subsapce

A generic function.}
\usage{
print.subspace(x, ...)
}
\arguments{
\item{x}{a subspace class.}

\item{...}{Extra parameters}
}
\description{
Print subsapce

A generic function.
}
