% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{roots_cubic}
\alias{roots_cubic}
\title{roots_cubic}
\usage{
roots_cubic(a2, a1, a0)
}
\arguments{
\item{a2}{An \tt{R} object.}

\item{a1}{An \tt{R} object.}

\item{a0}{An \tt{R} object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
