% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{roots_quartic}
\alias{roots_quartic}
\title{roots_quartic}
\usage{
roots_quartic(a3, a2, a1, a0)
}
\arguments{
\item{a3}{An R object.}

\item{a2}{An R object.}

\item{a1}{An R object.}

\item{a0}{An R object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
