% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{ruv_A}
\alias{ruv_A}
\title{ruv_A}
\usage{
ruv_A(A, nsamp = 25)
}
\arguments{
\item{A}{An R object.}

\item{nsamp}{An R object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
