% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreePlot.R
\name{ScreePlot}
\alias{ScreePlot}
\title{Scree Plot of scaled eigenvalues of X and sampled Expected Eigenvalues from Marcenko-Pastur distribution}
\usage{
ScreePlot(MPSamples, Changepoint = NULL, annotation = NULL)
}
\arguments{
\item{X}{A list. Samples from Marc\u{e}nko-Pastur (MP) distribution and eigenvalues of X. Output from MarcenkoPasturSample function.}
}
\description{
This function plot scree plot of scaled eigenvalues of X and sampled scaled expected eigenvalues from Marcenko-Pastur distribution.
}
\examples{
\donttest{
ScreePlot(MPSamples,Changepoint=2,annotation=0)
}
}
