% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_estimator_code.R
\name{slicing}
\alias{slicing}
\title{slicing}
\usage{
slicing(y, h, n)
}
\arguments{
\item{y}{numbers.}

\item{h}{numbers.}

\item{n}{numbers.}
}
\description{
All credit to Luo, W. and Li, B. (2016),
Combining Eigenvalues and Variation of Eigenvectors for Order
Determination, Biometrika, 103, 875–887. <doi:10.1093/biomet/asw051>
}
