% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{subspace.default}
\alias{subspace.default}
\title{A constructor function for the subspace class}
\usage{
subspace.default(
  x,
  components,
  decomposition,
  mp,
  num_est_samples,
  verbose,
  ...
)
}
\arguments{
\item{x}{A numeric real-valued matrix with n number of samples and
p number of features. If p > n, a warning message is generated and
the transpose of x is used.}

\item{components}{A series of right singular vectors to estimate.
Components must be smaller or equal to min(nrow(x), ncol(x)).}

\item{decomposition}{The method to be used; method = "svd"
returns results from singular value decomposition; method = "eigen"
returns results from eigenvalue decomposition.}

\item{mp}{A logical value. If true, sample eigenvlaues from random noise
matrix with mp distribution.}

\item{num_est_samples}{Split data into num_est_samples-fold for
parallel computation.}

\item{verbose}{output message}

\item{...}{Extra parameters}
}
\description{
This function calculates scaled eigenvalues
 and eigenvectors of x matrix, as well as sampled eigenvalues
 from a random noise matrix N of the same dimension as x,
 which follows a Marcenko-Pastur distribution with package
 "RMTsata"(https://cran.r-project.org/web/packages/RMTstat/index.html).
}
\seealso{
* [MarchenkoPasturPar()] calculates upper and lower limits
 of Marcenko-Pastur distribution from RMTstat package.

* [rmp()] sample scaled eigenvalues of random noise matrix
 from RMTstat package.
}
