% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{subspace}
\alias{subspace}
\title{a constructor function for the "subspace" class}
\usage{
subspace(X = X, rank = NA, basis = c("eigen", "singular"),
  times = NA, verbose = TRUE, ...)
}
\arguments{
\item{X}{A numeric real- or real-valued sparse matrix with n number of samples and p number of features. If p>n, a warning message is generated and the transpose of X is used.}

\item{rank}{a series of right singular vectors to estimate. rnk must be smaller or equal to max(nrow(X),ncol(X)).}

\item{basis}{??????????????????????}

\item{times}{split data into X times for parallel computation.}
}
\value{
Returns a list with entries:
\describe{
  \item{ndf:}{ number of degrees of freedom of X and the white Wishart matrix.}
  \item{pdim:}{ number of dimensions of X and the white Wishart matrix.}
  \item{rank:}{ a series of right singular vectors estimated.}
  \item{var_correct:}{ population variance for Marcenko-Pastur distribution.}
  \item{transpose_flag:}{ whether the matrix X is transposed.}
  \item{irl:}{ a data frame of scaled eigenvalues and corresponding dimensions.}
  \item{MP_irl:}{ a data frame of samped expected eigenvalues from Marcenko-Pastur and corresponding dimensions.}
  \item{v:}{ right singular vectors of X matrix with truncation up to dimension rnk.}
  \item{u:}{ left singular vectors of X matrix with truncation up to dimension rnk.}
}
}
\description{
This function sample scaled expected eigenvalues from Marcenko-Pastur distribution with package "RMTsata"(https://cran.r-project.org/web/packages/RMTstat/index.html).
}
\examples{
\donttest{
X <- Xsim(n = 1000, p = 500, ncc = 10, var = 2, fact = 1)
subspace_ <- subspace(X, rnk = 40, times = 10, basis = "eigen")
}
should import RMTstat after fix bug
}
