% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Xsim}
\alias{Xsim}
\title{simulate X matrix}
\usage{
Xsim(n = 100, p = 90, ncc = 2, var = 10, orthogonl = FALSE,
  seed = 1000)
}
\arguments{
\item{n}{number of rows for matrix X.}

\item{p}{number of columns for matrix X.}

\item{ncc}{number of correlated columns.}

\item{orthogonl}{a logical value. Generate orthogonal noise if true.}

\item{seed}{the random number seed.}

\item{sigma}{covariance matrix.}
}
\description{
These functions provide multivariate normal distributed X matrix.
}
\examples{
\donttest{
# 3 baskets, each with enrollement size 5
X <- Xsim(n=100,p=90,ncc=2,var=10)
}
}
