% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDimMatrix.R
\name{CheckDimMatrix}
\alias{CheckDimMatrix}
\title{Return parameter setting for MarcenkoPasturSample function.}
\usage{
CheckDimMatrix(X = X, rnk = NA, verbose = TRUE)
}
\arguments{
\item{X}{A numeric real- or real-valued sparse matrix with n number of samples and p number of features. If p>n, a warning message is generated and the transpose of X is used.}

\item{rnk}{number of right singular vectors to estimate. rnk must be smaller or equal to max(nrow(X),ncol(X)).}
}
\value{
Returns a list with entries:
\describe{
  \item{ndf:}{ number of degrees of freedom of X and the white Wishart matrix.}
  \item{pdim:}{ number of dimensions of X and the white Wishart matrix.}
  \item{svr:}{ samples to variance ratio; equals to ndf divided by pdim.}
  \item{rnk:}{ number of right singular vectors to estimate.}
  \item{transpose_flag:}{ whether the matrix X is transposed.}
}
}
\description{
Return parameter setting for MarcenkoPasturSample function.
}
\examples{
\donttest{
X <- Xsim(n=1000,p=500,ncc=10,var=2,fact = 1)
params <- CheckDimMatrix(X,rnk=40)
}
#should import RMTstat after fix bug
}
\seealso{
Random Matrix Theory pacakge credit to Gregory Giecold and Lionel Ouaknin
}
