% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R
\name{dimension}
\alias{dimension}
\title{Rank Estimation in High-Dimensional matrix X.}
\usage{
dimension(X = X, subspace_ = NULL, rank = NA, basis = c("eigen",
  "singular"), times = NA, p = 0.9, verbose = TRUE, ...)
}
\arguments{
\item{X}{A numeric real- or real-valued sparse matrix with n number of samples and p number of features. If p > n, a warning message is generated and the transpose of X is used.}

\item{times}{split data into X times for parallel computation.}

\item{p}{threshold for selecting changepoint.default is 0.90.}

\item{MPSamples}{A list. Samples from Marc\u{e}nko-Pastur (MP) distribution and eigenvalues of X. Output from MarcenkoPasturSample function.}

\item{rnk}{number of right singular vectors to estimate. rnk must be smaller or equal to max(nrow(X),ncol(X)).}
}
\value{
Returns a list with entries:
\describe{
  \item{ndf:}{ number of degrees of freedom of X and the white Wishart matrix.}
  \item{pdim:}{ number of dimensions of X and the white Wishart matrix.}
  \item{rnk:}{ number of right singular vectors to estimate.}
  \item{var_correct:}{ population variance for Marcenko-Pastur distribution.}
  \item{transpose_flag:}{ whether the matrix X is transposed.}
  \item{irl:}{ a data frame of scaled eigenvalues and corresponding dimensions.}
  \item{MP_irl:}{ a data frame of samped expected eigenvalues from Marcenko-Pastur and corresponding dimensions.}
  \item{v:}{ right singular vectors of X matrix with truncation up to dimension rnk.}
  \item{u:}{ left singular vectors of X matrix with truncation up to dimension rnk.}
  \item{bcp_irl:}{ probability of change in mean and posterior means of eigenvalue difference between $X$ and $N$.}
  \item{bcp_post:}{ probability of change in mean and posterior means of bcp_rirl.}
  \item{changePoint:}{ estimated changepoint position by from bcp_post.}
}
}
\description{
Estimate a low rank approximation of a signal-rich subspace in large high-dimensional data.
}
\section{Details}{

We estimate a low rank approximation of a signal-rich subspace in large high-dimensional data by decomposing matrix
into a signal-plus-noise space and approximate the signal-rich subspace with a rank K approximation
\eqn{\hat{X}=\sum_{k=1}^{K}d_ku_k{v_k}^T}. To estimate rank K, we propose a simple procedure assuming that matrix X is composed
of a low-rank signal matrix S and an average general noise random matrix \eqn{\bar{N}}. It has been shown that
the average eigenvalues of random matrices N follows a universal Marc\u{e}nko-Pastur (MP) distribution.
We hypothesize that the deviation of eigenvalues of X from the MP distribution indicates the intrinsic dimension of signal-rich subspace.
}

\examples{
\donttest{
X <- Xsim(n=1000,p=500,ncc=10,var=2,fact = 1,orthogonl = FALSE)
results<-dimension(X,rnk=10,times=100)

#equivelantly, if MarcenkoPasturSample is calcualted
MPSamples<-MarcenkoPasturSample(X,rnk=40,times=100)
results = dimension(MPSamples=MPSamples)
str(results)
ScreePlot(results$MarcenkoPasturSample,Changepoint=results$Changepoint$changePoint,annotation=10)
modified_legacyplot(results$Changepoint$bcp_post,annotation=10)
}
should import RMTstat after fix bug
}
\seealso{
\code{\link[RMTstat]} for details of Marcenko-Pastur distribution.
}
