% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff.R
\name{hoff_gibbs_UVD_varrank}
\alias{hoff_gibbs_UVD_varrank}
\title{hoff_gibbs_UVD_varrank}
\usage{
hoff_gibbs_UVD_varrank(U, V, D, Y, phi, mu, psi, nsamp)
}
\arguments{
\item{U}{An \tt{R} object.}

\item{V}{An \tt{R} object.}

\item{D}{An \tt{R} object.}

\item{Y}{An \tt{R} object.}

\item{phi}{An \tt{R} object.}

\item{mu}{An \tt{R} object.}

\item{psi}{An \tt{R} object.}

\item{nsamp}{An \tt{R} object.}
}
\description{
All credit to Hoff, Peter D. "Model averaging
and dimension selection for the singular value decomposition."
Journal of the American Statistical Association 102.478 (2007): 674-685.
}
