% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_k_means.R
\name{km}
\alias{km}
\title{Modified K means clustering}
\usage{
km(data)
}
\arguments{
\item{data}{an n by p numeric matrix; the data
matrix of predictors}
}
\value{
a data frame of total within-cluster variation.
}
\description{
k-means clustering modified to cluster probabilities of change points
to 2 clusters and return total within-cluster variation. Adapted from
https://github.com/TunChiehHsu/Kmeans.git.
}
