% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung.R
\docType{data}
\name{lung}
\alias{lung}
\title{IPF Single Cell Altas data}
\format{
An object of class Seurat.
}
\source{
\href{https://www.biorxiv.org/content/10.1101/759902v1.full.pdf}
{https://www.biorxiv.org/content/10.1101/759902v1.full.pdf}
}
\usage{
data(lung)
}
\description{
The lung data is a subset of patient "001C" profiled in IPF
Single Cell Altas. The full data seet can be found on GEO (GSE136831),
which was examined  by
\href{https://www.biorxiv.org/content/10.1101/759902v1.full.pdf}
{https://www.biorxiv.org/content/10.1101/759902v1.full.pdf}
to build a single cell atlas of Idiopathic Pulmonary Fibrosis (IPF).
It contains the gene expression of 312,928 cells profiled
from 32 IPF patient lung samples, 18 chronic obstructive pulmonary
disease (COPD) lung samples and 28 control donor lung samples.
(Adamset al.2019).
}
\examples{
\dontrun{
data(lung)
}
}
\references{
Adamset  al.(2019),
(\href{https://www.biorxiv.org/content/10.1101/759902v1.full.pdf}
{https://www.biorxiv.org/content/10.1101/759902v1.full.pdf})
}
\keyword{datasets}
