% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.subspace.R
\name{plot.subspace}
\alias{plot.subspace}
\title{Scree plot of scaled eigenvalues of X and random noise matrix N}
\usage{
\method{plot}{subspace}(obj, Changepoint = NULL, annotation = NULL,
  verbose = TRUE)
}
\arguments{
\item{Changepoint}{A number. Estimated changepoint in OptimumDimension function.}

\item{annotation}{A number. Choose to label points up to annotation number. Set to 0 with no annotation.}

\item{data}{A subsapce class.}
}
\description{
This is a generic function for supspace class to plot scree plot of scaled eigenvalues of X and sampled scaled expected eigenvalues from Marcenko-Pastur distribution.
}
\examples{
\donttest{
plot(Subspace, Changepoint = 0, annotation = 15)
}
}
