% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R,
%   R/estimate_rank_double_posterior.R, R/estimate_rank_double_posterior_cor.R,
%   R/estimate_rank_kmeans.R, R/estimate_rank_kmeans_cor.R,
%   R/estimate_rank_posterior.R, R/estimate_rank_posterior_cor.R
\name{print.dimension}
\alias{print.dimension}
\title{Print dimension

A generic function.}
\usage{
print.dimension(x, ...)

print.dimension(x, ...)

print.dimension(x, ...)

print.dimension(x, ...)

print.dimension(x, ...)

print.dimension(x, ...)

print.dimension(x, ...)
}
\arguments{
\item{x}{a dimension class.}

\item{...}{Extra parameters}
}
\description{
Print dimension

A generic function.

Print dimension

A generic function.

Print dimension

A generic function.

Print dimension

A generic function.

Print dimension

A generic function.

Print dimension

A generic function.

Print dimension

A generic function.
}
