% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{subspace}
\alias{subspace}
\title{A constructor function for the subspace class}
\usage{
subspace(x, components = NA, mp = TRUE, times = NA, verbose = TRUE, ...)
}
\arguments{
\item{x}{A numeric real-valued matrix with n number of samples and
p number of features. If p > n, a warning message is generated and
the transpose of x is used.}

\item{components}{A series of right singular vectors to estimate.
Components must be smaller or equal to min(nrow(x), ncol(x)).}

\item{mp}{A logical value. If true, sample eigenvlaues from random noise
matrix with mp distribution.}

\item{times}{Split data into times-fold for parallel computation.}

\item{verbose}{output message}

\item{...}{Extra parameters}
}
\value{
Returns a list with entries:
\describe{
  \item{ndf:}{ The number of degrees of freedom of x.}
  \item{pdim:}{ The number of dimensions of x.}
  \item{components:}{ A series of right singular
   vectors estimated.}
  \item{var_correct:}{ Corrected population variance
   for Marcenko-Pastur distribution.}
  \item{transpose_flag:}{ A logical value indicating
   whether the matrix x is transposed.}
  \item{irl:}{ A data frame of scaled eigenvalues for
   specified components and corresponding dimensions.}
  \item{sigma_a:}{ A vector of scaled eigenvalues up to max(components).}
  \item{mp_irl:}{ A data frame of sampled expected eigenvalues from
   Marcenko-Pastur for specified components and corresponding dimensions.}
  \item{sigma_mp:}{ A vector of samped expected eigenvalues from
   Marcenko-Pastur up to max(components).}
  \item{v:}{ Right singular vectors of x matrix for specified components.}
  \item{u:}{ Left singular vectors of x matrix or specified components.}
}
}
\description{
This function calculates scaled eigenvalues
 and eigenvectors of x matrix, as well as sampled eigenvalues
 from a random noise matrix N of the same dimension as x,
 which follows a Marcenko-Pastur distribution with package
 "RMTsata"(https://cran.r-project.org/web/packages/RMTstat/index.html).
}
\examples{
x <- x_sim(n = 100, p = 150, ncc = 10, var = c(rep(10,5),rep(1,5)))
Subspace <- subspace(x)
Subspace <- subspace(x, components = 8:30)
Subspace <- subspace(x, components = c(2, 3, 6, 16))
Subspace <- subspace(x, components = 1:20, mp= FALSE)
Subspace
plot(Subspace, changepoint = 0, annotation = 1:10)
}
\seealso{
* [MarchenkoPasturPar()] calculates upper and lower limits
 of Marcenko-Pastur distribution from RMTstat package.

* [rmp()] sample scaled eigenvalues of random noise matrix
 from RMTstat package.
}
