% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_sim.R
\name{x_sim}
\alias{x_sim}
\title{Simulate x matrix}
\usage{
x_sim(n = 100, p = 90, ncc = 2, var = 10, seed = 1000)
}
\arguments{
\item{n}{The number of rows for matrix x.}

\item{p}{The number of columns for matrix x.}

\item{ncc}{The number of correlated columns.}

\item{var}{Varaince.}

\item{seed}{The random number seed.}
}
\description{
These functions provide multivariate normal
 distributed x matrix.
}
\examples{
\donttest{
x <- x_sim(n = 150, p = 100, ncc = 10, var = 2)
}
}
