% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Xsim}
\alias{Xsim}
\title{simulate X matrix}
\usage{
Xsim(n = 100, p = 90, ncc = 2, var = 10, fact = 30,
  seed = 1000)
}
\arguments{
\item{n}{number of rows for matrix X.}

\item{p}{number of columns for matrix X.}

\item{ncc}{number of correlated columns.}

\item{var}{varaince.}

\item{fact}{scale on basis if orthogonal is true.}

\item{seed}{the random number seed.}
}
\description{
These functions provide multivariate normal distributed X matrix.
}
\examples{
\donttest{
# 3 baskets, each with enrollement size 5
X <- Xsim(n=100,p=90,ncc=2,var=10,fact = 30)
}
}
