% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xsim.R
\name{Xsim}
\alias{Xsim}
\title{Simulate X matrix}
\usage{
Xsim(n = 100, p = 90, ncc = 2, var = 10, seed = 1000)
}
\arguments{
\item{n}{The number of rows for matrix X.}

\item{p}{The number of columns for matrix X.}

\item{ncc}{The number of correlated columns.}

\item{var}{Varaince.}

\item{seed}{The random number seed.}
}
\description{
These functions provide multivariate normal distributed X matrix.
}
\examples{
\donttest{
# 3 baskets, each with enrollement size 5
X <- Xsim(n = 150, p = 100, ncc = 10, var = 2)
}
}
